refactor recorder: extract controls into separate widget
