+++ open_wearable/lib/apps/widgets/apps_page.dart
@@ -1,14 +1,16 @@
+import 'dart:io';
+
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
 import 'package:go_router/go_router.dart';
 import 'package:open_earable_flutter/open_earable_flutter.dart';
+import 'package:open_wearable/apps/audio_recorder/view/audio_recorder_view.dart';
 import 'package:open_wearable/apps/heart_tracker/widgets/heart_tracker_page.dart';
 import 'package:open_wearable/apps/posture_tracker/model/earable_attitude_tracker.dart';
 import 'package:open_wearable/apps/posture_tracker/view/posture_tracker_view.dart';
 import 'package:open_wearable/apps/widgets/select_earable_view.dart';
 import 'package:open_wearable/apps/widgets/app_tile.dart';
-
 
 class AppInfo {
   final String logoPath;
@@ -29,15 +31,17 @@ List<AppInfo> _apps = [
     logoPath: "lib/apps/posture_tracker/assets/logo.png",
     title: "Posture Tracker",
     description: "Get feedback on bad posture",
-    widget: SelectEarableView(startApp: (wearable, sensorConfigProvider) {
-      return PostureTrackerView(
-        EarableAttitudeTracker(
-          wearable as SensorManager,
-          sensorConfigProvider,
-          wearable.name.endsWith("L"),
-        ),
-      );
-    },),
+    widget: SelectEarableView(
+      startApp: (wearable, sensorConfigProvider) {
+        return PostureTrackerView(
+          EarableAttitudeTracker(
+            wearable as SensorManager,
+            sensorConfigProvider,
+            wearable.name.endsWith("L"),
+          ),
+        );
+      },
+    ),
   ),
   AppInfo(
     logoPath: "lib/apps/heart_tracker/assets/logo.png",
@@ -48,8 +52,10 @@ List<AppInfo> _apps = [
         if (wearable is SensorManager) {
           //TODO: show alert if no ppg sensor is found
           Sensor ppgSensor = (wearable as SensorManager).sensors.firstWhere(
-            (s) => s.sensorName.toLowerCase() == "photoplethysmography".toLowerCase(),
-          );
+                (s) =>
+                    s.sensorName.toLowerCase() ==
+                    "photoplethysmography".toLowerCase(),
+              );
 
           return HeartTrackerPage(ppgSensor: ppgSensor);
         }
@@ -64,6 +70,13 @@ List<AppInfo> _apps = [
       },
     ),
   ),
+  if (Platform.isAndroid)
+    AppInfo(
+      logoPath: "lib/apps/audio_recorder/assets/logo.png",
+      title: "Audio Recorder",
+      description: "Record system audio and Bluetooth streams",
+      widget: const AudioRecorderView(), // Your audio recorder page
+    ),
 ];
 
 class AppsPage extends StatelessWidget {
@@ -75,7 +88,7 @@ class AppsPage extends StatelessWidget {
       appBar: PlatformAppBar(
         title: PlatformText("Apps"),
         trailingActions: [
-            PlatformIconButton(
+          PlatformIconButton(
             icon: Icon(context.platformIcons.bluetooth),
             onPressed: () {
               context.push('/connect-devices');
