+++ open_wearable/pubspec.yaml
@@ -54,6 +54,7 @@ dependencies:
   shared_preferences: ^2.5.3
   url_launcher: ^6.3.2
   go_router: ^14.6.2
+  playback_capture: ^0.0.4
 
 dev_dependencies:
   flutter_test:
